+++ pages/hobbies/index.vue
@@ -1,10 +1,13 @@
 <template>
   <ComingSoon v-if="!page">{{ $t('hobbies.title') }}</ComingSoon>
-  <div v-else class="section-experiences">
+  <div v-else class="section-hobbies">
     <div class="container-page">
-      <header class="section-experiences__title">
+      <header class="section-hobbies__title">
         <BaseHeading primary :label="$t('hobbies.subtitle')">{{ $t('hobbies.title') }}</BaseHeading>
       </header>
+
+      <StoryblokComponent v-for="blok in bloks" :key="blok._uid" :blok="blok" class="section-hobbies__blok" />
+
       <EndOfPage />
     </div>
   </div>
@@ -22,9 +25,23 @@ const { $setMetadata } = useNuxtApp()
 
 const { getParams } = useLocalizedStoryParams()
 
-const page = await useAsyncStoryblok('hobbies', getParams()).catch(() => null)
+const page = await useAsyncStoryblok('hobbies', getParams())
+  .then((data) => data.value.content)
+  .catch(() => null)
+
+const bloks = computed(() => {
+  return page.body
+})
 
 $setMetadata({
   title: t('hobbies.title'),
 })
 </script>
+
+<style lang="scss">
+.section-hobbies {
+  &__blok {
+    margin-bottom: 2rem;
+  }
+}
+</style>