+++ storyblok/LongText.vue
@@ -16,5 +16,10 @@ defineProps<{
 
 <style lang="scss">
 .long-text {
+  &__title {
+    font-weight: bold;
+    font-size: 2.5rem;
+    color: var(--color-neutral-darker);
+  }
 }
 </style>